Add types to makeTemporaryEnv
