+++ packages/acceptance-tests/pkg-tests-core/sources/index.ts
@@ -1,3 +1,5 @@
+/// <reference path="../types/global.d.ts" />
+
 import * as exec  from './utils/exec';
 import * as fs    from './utils/fs';
 import * as misc  from './utils/misc';
